% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risco_test_1.R
\docType{data}
\name{risco_test_1}
\alias{risco_test_1}
\title{ISCO88 risco testdata 1}
\format{Data frame with 510 rows and 2 columns
\describe{
  \item{isco_code}{ISCO88 code of occupations}
  \item{english}{ISCO88 labels in english}
  ...
}}
\source{
\url{http://www2.warwick.ac.uk/fac/soc/ier/research/classification/isco88/}
}
\usage{
risco_test_1
}
\description{
ISCO88 codes used for examples.
}
\keyword{datasets}
