% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risco_test_2.R
\docType{data}
\name{risco_test_2}
\alias{risco_test_2}
\title{ISCO88 risco testdata 2}
\format{Data frame with 10 rows and 2 columns
\describe{
  \item{isco_code}{ISCO88 code of occupations}
  \item{english}{ISCO88 labels in english}
  ...
}}
\source{
\url{http://www2.warwick.ac.uk/fac/soc/ier/research/classification/isco88/}
}
\usage{
risco_test_2
}
\description{
10 ISCO88 codes with two missings for example.
}
\keyword{datasets}
