% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco_relabel.R
\name{isco_relabel}
\alias{isco_relabel}
\title{ISCO relabeling in different languages}
\usage{
isco_relabel(isco_col, label_col, dest_language = "english")
}
\arguments{
\item{isco_col}{Column in where the ISCO codes(2 to 4 digits) are stored. Please indicate
via subsetting.}

\item{label_col}{Column where the isco labels are stored. Please indicate via subsetting.}

\item{dest_language}{Denstination language. Indicate on of "german", "english", "french"
as a string. Defaults to "english".}
}
\description{
Thisfunction takes isco88 codes (2 to 4 digits) as its argument and changes the
labels in the specified column (second argument) to the desired language.
}
