% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\docType{data}
\name{labels}
\alias{labels}
\title{ISCO 88 labels in English, French and German}
\format{Data frame with 510 rows and 4 columns
\describe{
  \item{isco_code}{ISCO 88 code of occupations}
  \item{english}{ISCO 88 labels in english}
  \item{french}{ISCO 88 labels in french}
  \item{german}{ISCO 88 labels in german}
  ...
}}
\source{
\url{http://www2.warwick.ac.uk/fac/soc/ier/research/classification/isco88/}
}
\usage{
labels
}
\description{
Dataset from the Warwick Institute for Employment Research containing ISCO88 labels.
}
\keyword{datasets}
